+++ src/shared/ui/button/index.ts
@@ -1,2 +1,2 @@
-export { default as Button } from "./button";
+export * from "./button";
 export { default as SelectedIconButton } from "./selectedIconButton";
